# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface PlayerBalance {
  /**
     * @minLength 1
     * @maxLength 80
     */
  name: string;
  /** @minimum 0 */
  endingAmount: number;
  /** @minimum 0 */
  zhaHuCount: number;
  /** @minimum 0 */
  xieXieKaiXiangCount: number;
}

export interface MahjongSession {
  id: number;
  playedOn: string;
  /** @minimum 1 */
  rounds: number;
  totalAmount: number;
  winnerName: string;
  playerBalances: PlayerBalance[];
  /** @nullable */
  notes: string | null;
  /** @nullable */
  createdByUserId: string | null;
  createdAt: string;
}

export interface MahjongSessionInput {
  playedOn: string;
  /**
     * @minimum 1
     * @maximum 99
     */
  rounds: number;
  /**
     * @minItems 4
     * @maxItems 4
     */
  playerBalances: PlayerBalance[];
  /**
     * @maxLength 500
     * @nullable
     */
  notes?: string | null;
}

export interface MahjongSessionUpdate {
  playedOn?: string;
  /**
     * @minimum 1
     * @maximum 99
     */
  rounds?: number;
  /**
     * @minItems 4
     * @maxItems 4
     */
  playerBalances?: PlayerBalance[];
  /**
     * @maxLength 500
     * @nullable
     */
  notes?: string | null;
}

export type ChipDenomination = typeof ChipDenomination[keyof typeof ChipDenomination];


export const ChipDenomination = {
  white: 'white',
  orange: 'orange',
  light_blue: 'light_blue',
  blue: 'blue',
  black: 'black',
} as const;

export type ChipScanInputMimeType = typeof ChipScanInputMimeType[keyof typeof ChipScanInputMimeType];


export const ChipScanInputMimeType = {
  'image/jpeg': 'image/jpeg',
  'image/png': 'image/png',
  'image/webp': 'image/webp',
} as const;

export interface ChipScanInput {
  /**
     * Base64-encoded image bytes without a data URL prefix.
     * @minLength 16
     * @maxLength 6000000
     */
  imageBase64: string;
  mimeType: ChipScanInputMimeType;
}

export interface ChipStackEstimate {
  denomination: ChipDenomination;
  /**
     * @minimum 1
     * @maximum 500
     */
  count: number;
  /**
     * @minimum 0
     * @maximum 1
     */
  confidence: number;
  /**
     * @maxLength 240
     * @nullable
     */
  uncertainty: string | null;
}

export interface ChipScanResult {
  /**
     * @minItems 1
     * @maxItems 5
     */
  stacks: ChipStackEstimate[];
  /**
     * @minimum 0
     * @maximum 1
     */
  overallConfidence: number;
  /**
     * @maxLength 300
     * @nullable
     */
  guidance: string | null;
}

export interface WinnerCount {
  winnerName: string;
  wins: number;
}

export interface ZhaHuCount {
  playerName: string;
  /** @minimum 0 */
  count: number;
}

export interface XieXieKaiXiangCount {
  playerName: string;
  /** @minimum 0 */
  count: number;
}

export interface PlayerWinnings {
  playerName: string;
  netAmount: number;
}

export interface SessionSummary {
  totalSessions: number;
  totalRounds: number;
  totalAmount: number;
  latestSession: MahjongSession | null;
  winnerCounts: WinnerCount[];
  zhaHuCounts: ZhaHuCount[];
  xieXieKaiXiangCounts: XieXieKaiXiangCount[];
  playerWinnings: PlayerWinnings[];
}