# lib/api-zod/src/generated/types/chipDenomination.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export type ChipDenomination = typeof ChipDenomination[keyof typeof ChipDenomination];


export const ChipDenomination = {
  white: 'white',
  orange: 'orange',
  light_blue: 'light_blue',
  blue: 'blue',
  black: 'black',
} as const;