# lib/api-zod/src/generated/types/chipScanInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ChipScanInputMimeType } from './chipScanInputMimeType';

export interface ChipScanInput {
  /**
     * Base64-encoded image bytes without a data URL prefix.
     * @minLength 16
     * @maxLength 6000000
     */
  imageBase64: string;
  mimeType: ChipScanInputMimeType;
}