# lib/api-zod/src/generated/types/chipScanResult.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ChipStackEstimate } from './chipStackEstimate';

export interface ChipScanResult {
  /**
     * @minItems 1
     * @maxItems 5
     */
  stacks: ChipStackEstimate[];
  /**
     * @minimum 0
     * @maximum 1
     */
  overallConfidence: number;
  /**
     * @maxLength 300
     * @nullable
     */
  guidance: string | null;
}