# lib/api-zod/src/generated/types/chipStackEstimate.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ChipDenomination } from './chipDenomination';

export interface ChipStackEstimate {
  denomination: ChipDenomination;
  /**
     * @minimum 1
     * @maximum 500
     */
  count: number;
  /**
     * @minimum 0
     * @maximum 1
     */
  confidence: number;
  /**
     * @maxLength 240
     * @nullable
     */
  uncertainty: string | null;
}